requirements.txt
httpx~=0.28.1
fastapi~=0.115.14
sqlmodel~=0.0.24
python-dotenv~=1.1.1
pydantic~=2.11.7